Transform apply_credential into a HTTP policy
